Install missing utilities (#2)
* Add CM5 device tree files

This is untested, but should make it more likely for CM5's to work when
people start getting them.

* Include support for the Hailo AI accelerator

This only enables Linux kernel support.

* nerves_system_br: bump to v1.29.2

* nerves_system_br: bump to v1.29.3

* Bump Linux to 6.6.64; add new cm5 device tree files

* Enable RP1 PIO driver and add notes

This will enable some pretty cool I/O on the RPi5 for those interested
in programming the PIO state machines.

* Add missing overlay needed for the CM5

This was found by running `vclog --msg` on a booting image and seeing
that `bcm2712d0.dtbo` was being loaded. The docs says that this overlay
has the differences from the c0 silicon that seems to be on all of the
regular RPi 5's right now.

* Support NVME drives now that M.2 slots are common

* Enable libdrm for libcamera preview

* Resave nerves_defconfig: libdrm enabled by default

* v0.4.1 release

* Install missing utilities

---------

Co-authored-by: Frank Hunleth <fhunleth@troodon-software.com>
Co-authored-by: Damirados <damirados@gmail.com>